+++ grunt/options/uglify.js
@@ -1,5 +1,5 @@
 //
-// Use to uglify JS
+// Used to uglify JS
 //
 module.exports = {
 	dist: {
@@ -8,7 +8,8 @@ module.exports = {
 			'dist/scripts/modules/swagger-external-references.min.js': ['dist/scripts/modules/swagger-external-references.js'],
 			'dist/scripts/modules/swagger-xml-formatter.min.js': ['dist/scripts/modules/swagger-xml-formatter.js'],
 			'dist/scripts/modules/swagger1-to-swagger2-converter.min.js': ['dist/scripts/modules/swagger1-to-swagger2-converter.js'],
-			'dist/scripts/modules/swagger-yaml-parser.min.js': ['dist/scripts/modules/swagger-yaml-parser.js']
+			'dist/scripts/modules/swagger-yaml-parser.min.js': ['dist/scripts/modules/swagger-yaml-parser.js'],
+			'dist/scripts/i18n/fr.min.js': ['dist/scripts/i18n/fr.js']
 		}
 	}
 }